Teste3/requirements.txt
SQLAlchemy
pandas
python-dotenv
requests
psycopg2